requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ML
lightgbm>=4.0.0
scikit-learn>=1.3.0
shap>=0.42.0

# Data
pyarrow>=13.0.0

# Zerodha
kiteconnect>=5.0.0

# Config & Logging
pyyaml>=6.0.0
python-json-logger>=2.0.0

# Visualization
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
